# [MRG] PERF Significant performance improvements in Partial Least Squares (PLS) Regression