--- a/sklearn/cross_decomposition/_pls.py
+++ b/sklearn/cross_decomposition/_pls.py
@@ -695,14 +695,14 @@ class PLSRegression(_PLS):
     #     - "pls" with function oscorespls.fit(X, Y)
 
     def __init__(
-        self, n_components=2, *, scale=True, max_iter=500, tol=1e-06, copy=True
+        self, n_components=2, *, scale=True, max_iter=500, tol=1e-06, copy=True, algorithm="nipals"
     ):
         super().__init__(
             n_components=n_components,
             scale=scale,
             deflation_mode="regression",
             mode="A",
-            algorithm="nipals",
+            algorithm=algorithm,
             max_iter=max_iter,
             tol=tol,
             copy=copy,