--- a/sklearn/cross_decomposition/_pls.py
+++ b/sklearn/cross_decomposition/_pls.py
@@ -385,7 +385,7 @@ class _PLS(
                 P[:, k] = p
                 Q[:, [k]] = q
 
-            self._coef_ = (R @ Q.T) * self._y_std
+            self._coef_ = (R @ Q.T).T * self._y_std
             self.intercept = self._y_mean
 
         elif self.algorithm == "simpls":
@@ -422,7 +422,7 @@ class _PLS(
                 U[:, [k]] = u
                 V[:, [k]] = v
 
-            self._coef_ = (R @ Q.T) * self._y_std
+            self._coef_ = (R @ Q.T).T * self._y_std
             self.intercept = self._y_mean
 
         return self