… doc/whats_new/v1.2.rst
… examples/cross_decomposition/plot_compare_cross_decomposition.py
… examples/cross_decomposition/plot_pcr_vs_pls.py
… sklearn/cross_decomposition/tests/test_pls.py