@fractionalhare 2022-07-10
I'm not sure why the Sphinx documentation builder is failing - the additions to the `PLSRegression` docstring are around line 599, but I don't see the indentation error. What am I missing?