@fractionalhare 2022-07-13
Does this need a changelog entry?