@fractionalhare 2022-07-14
@ogrisel Given my results from walking through the two examples in the documentation, what's your take on keeping the default as NIPALS vs changing it to the new algorithm?