@ogrisel 2022-07-18
> Does this need a changelog entry?

yes it needs an entry of the "Enhancement" kind.