@ogrisel 2022-07-18
> @ogrisel Given my results from walking through the two examples in the documentation, what's your take on keeping the default as NIPALS vs changing it to the new algorithm?

I think that changing the default might require going through a `FutureWarning` cycle. I think we can try to finalize and review this PR without changing the default, and then discuss changing the default in a follow-up PR. I think we need to be very careful with evaluation when changing the default not to introduce any regressions.

In particular, are there cases where the previous default was more memory efficient, CPU efficient or more numerically stable?

I am not familiar enough with PLS regression to answer this question, so I prefer to delay this discussion to another PR.