@fractionalhare 2022-07-18
> Can you try to rerun memit on a larger dataset?

Sure. With `n_samples = 10000` and `n_features = 2000`, `n_targets = 50` and `n_components = 50`:

NIPALS:

`peak memory: 646.89 MiB, increment: 380.25 MiB`

Dayal-MacGregor:

`peak memory: 718.62 MiB, increment: 71.73 MiB`