@ogrisel 2022-07-19
Thanks. Can you please enable the new solver in the example(s) that use `PLSRegressor` since it it faster and pushed this change as part of this PR? It should make documentation building a bit more efficient and will give more visibility to this new option.