@ogrisel 2022-07-20
Another remark, the sign of the extracted coef can sometimes flip when switch algorithm:

- https://scikit-learn.org/dev/auto_examples/cross_decomposition/plot_pcr_vs_pls.html
- https://output.circle-artifacts.com/output/job/334401fc-5576-4efe-bf6a-ec8aea0f37b8/artifacts/0/doc/auto_examples/cross_decomposition/plot_pcr_vs_pls.html

Would it be possible to use `sklearn.utils.extmath.svd_flip` or `_svd_flip_1d` to avoid this artifact? If not easy to achieve, no big deal.