@ogrisel 2022-08-12
> @ogrisel what is this waiting on?

Unfortunately the documentation building CI has timed out. Can you please try to push another commit (e.g. yet another merge with the current `main`) to see if this problem has been resolved? I am not sure about the cause.

It would be great to be able to checkout the rendered HTML for the examples with your latest changes.