@ogrisel 2022-08-12
> This quirk is harmless for prediction because the coefficient matrix is x_rotations_ @ y_loadings_ so when the signs on individual column vectors are opposite due to SVD, the coefficient matrices still match.

I agree but the fact that the sign of the coef changes when switching solvers is practically annoying, e.g. when writing a doc with some explanation of a plot of the latent space: the top right part of the figure could silently become the bottom left of the figure if the code is switched from nipals to the new solver.

Since we don't plan to change the default right away this not necessarily a big problem as this subtle behavior change will not happen just by updating the version of scikit-learn, but still, I their had been a way to enforce this stability that would a nice convenience to our future users. See the [Principle of Least Astonishment](https://en.wikipedia.org/wiki/Principle_of_least_astonishment).