@lorentzenchr 2022-08-13
I'm not an expert of PLS, but it seems to me that the proposed Dayal-MacGregor modified kernel algorithm is superior to the current algo. Therefore, in the long run, I would not introduce a new `solver` parameter.
Due to the sign change, the open question is how to do a proper deprecation cycle. Could we introduce a new parameter `solver` which we deprecate right from the beginning, saying that the default will change and that this parameter will be removed in 2 versions?