@ogrisel 2022-08-16
The new solver does compute all the attributes that the nipals algorithm could compute. So I would still keep the solver attribute to be able to select the nipals algorithm shall we decide to move to Dayal-MacGregor as the default in a future version of scikit-learn.