@lorentzenchr 2022-08-17
Conclusion: Let's introduce the `solver` parameter

Possible (for me quite likely) future path outside of this PR: change default to the new solver, finally remove `solver` and get rid of the old solver.