@fractionalhare 2023-11-07
Last status was an initial review from @ogrisel - enough time has passed that the review should probably be at least refreshed, and then one other person needs to do a review.

I can update the branch and see what coverage checks are still failing; we had the CI all green before so it should just be a matter of update and then tweaking tests.